Windows Build Changes
For open sourcing purposes, this is the build to use pyinstaller to create the one file version of the youtube transcriber.

Whisper must be included, you must also have the required libraries installed as well (if you want to build it yourself, otherwise just download the .exe release)
Modifications to the original script were made in order to manage temp files properly as well as setting up for pyinstaller.

Use the build_command.txt for the pyinstaller command.